// sdk/ts/src/models/DailyDauResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DailyDauRow } from './DailyDauRow';
export type DailyDauResponse = {
    end: string;
    rows: Array<DailyDauRow>;
    start: string;
    timezone: string;
};